requirements.txt
azure-servicebus
numpy
